fix(backend): atualização para refletir as mudanças atuais
